% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pp}
\alias{plot_pp}
\title{P-P (probability-probability) plot}
\usage{
plot_pp(fit)
}
\arguments{
\item{fit}{A converged \code{"slt_fit"} object, fit to data from a
simple-step-stress (\code{\link[=rslt]{rslt()}} with \code{prop_transfer = 1}) design.}
}
\value{
A \code{ggplot} object.
}
\description{
P-P (probability-probability) plot
}
\examples{
set.seed(1)
d <- rslt(150, "weibull", c(0.05, 1.5), c(0.2, 1.5), tau1 = 5, prop_transfer = 1)
fit <- fit_slt(d, "weibull", "shape1", tau1 = 5)
plot_pp(fit)
}
