% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.slt_fit}
\alias{print.slt_fit}
\title{Print a fitted two-stage life testing model}
\usage{
\method{print}{slt_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{An \code{"slt_fit"} object from \code{\link[=fit_slt]{fit_slt()}}.}

\item{digits}{Number of digits to round to.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted two-stage life testing model
}
\seealso{
\code{\link[=summary.slt_fit]{summary.slt_fit()}} for a more detailed summary with
confidence limits.
}
