% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.slt_fit}
\alias{print.summary.slt_fit}
\title{Print a summary of a fitted two-stage life testing model}
\usage{
\method{print}{summary.slt_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{"summary.slt_fit"} object from \code{\link[=summary.slt_fit]{summary.slt_fit()}}.}

\item{digits}{Number of digits to round to.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a summary of a fitted two-stage life testing model
}
