% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_df}
\alias{round_df}
\title{Round a numeric vector/data.frame column set, silently ignoring non-numeric}
\usage{
round_df(df, digits = 4)
}
\description{
Round a numeric vector/data.frame column set, silently ignoring non-numeric
}
\keyword{internal}
